refactor(qtensor): move absmax_scale to calibration
