+++ quanto/tensor/core.py
@@ -12,14 +12,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from typing import Optional
 
 import torch
 
-from .qtype import qint8, qtype
 
-
-__all__ = ["absmax_scale", "axis_to_dim", "dtype_info", "group", "ungroup"]
+__all__ = ["axis_to_dim", "dtype_info", "group", "ungroup"]
 
 
 def dtype_info(dtype):
@@ -71,30 +68,3 @@ def ungroup(grouped: torch.Tensor, axis: int, orig_shape: torch.Size):
     # Permute to (axis_groups, group_size, axis_dim)
     ungrouped = ungrouped.permute(2, 0, 1)
     return ungrouped.reshape(orig_shape)
-
-
-def absmax_scale(base: torch.Tensor, qtype: qtype = qint8, axis: Optional[int] = None) -> torch.Tensor:
-    """Evaluate the quantization scale using the absmax algorithm.
-
-    The Absolute Maximum quantization algorithm is a symmetrical quantization
-    algorithm where the scale corresponds to the maximum absolute value of the
-    base divided by the highest positive integer value for the target integer
-    representation.
-
-    Args:
-        base (`torch.Tensor`): the base tensor on which the scale will be applied.
-        qtype (`quanto.qtype`): the target qtype for quantization.
-        axis (`int`): the index of the axis to preserve, or -1 for the last one.
-            Defaults to None to reduce all axis.
-
-    Returns:
-        `torch.Tensor`: a scale tensor of the same dtype as the base.
-    """
-    base = torch.abs(base)
-    if axis is None:
-        qranges = torch.max(base)
-    else:
-        dim = axis_to_dim(base, axis)
-        qranges = torch.amax(base, dim=dim, keepdim=True)
-    info = dtype_info(qtype.dtype)
-    return qranges / info.max
